Draft with updated "core-types"
This is not fully fleshed out on the Hydra.Direct.Tx level. It's unclear
whether that low-level module already could come up with whatever we
want to store as 'ChainState'. Rather the Hydra.Direct.State would do
that and the return types of 'observeAbortTx' et al should not directly
be 'OnChainTx' if that contains the 'ChainState'